Refactor admin server
