+++ admin/server.go
@@ -1,15 +1,9 @@
 package admin
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"log"
 	"net/http"
-
-	"github.com/visola/go-proxy/config"
-	myhttp "github.com/visola/go-proxy/http"
-	"github.com/visola/go-proxy/statistics"
 
 	"github.com/gobuffalo/packr"
 	"github.com/gorilla/mux"
@@ -47,40 +41,4 @@ func StartAdminServer() error {
 	adminServer.Handle("/{file:.*}", http.FileServer(box))
 
 	return http.ListenAndServe(fmt.Sprintf(":%d", adminServerPort), adminServer)
-}
-
-func handleConfigurations(w http.ResponseWriter, req *http.Request) {
-	configurations, configError := config.GetConfigurations()
-	if configError != nil {
-		myhttp.InternalError(req, w, configError)
-		return
-	}
-	responseWithJSON(configurations, w, req)
-}
-
-func handleRequets(w http.ResponseWriter, req *http.Request) {
-	if req.Header.Get("Upgrade") == "websocket" {
-		conn, err := upgrader.Upgrade(w, req, nil)
-		if err != nil {
-			log.Println(err)
-			return
-		}
-
-		statistics.OnRequestProxied(func(proxiedRequest statistics.ProxiedRequest) {
-			conn.WriteJSON(proxiedRequest)
-		})
-	} else {
-		responseWithJSON(statistics.GetProxiedRequests(), w, req)
-	}
-}
-
-func responseWithJSON(data interface{}, w http.ResponseWriter, req *http.Request) {
-	json, jsonError := json.Marshal(data)
-	if jsonError != nil {
-		myhttp.InternalError(req, w, jsonError)
-		return
-	}
-
-	w.Header().Add("Content-Type", "application/json")
-	w.Write([]byte(json))
 }
